enforcement of preferences selection (way of selection is to be changed), adding prefered project to user, simplify add tuple of messages function
changing function addMessagesTuplesOfSize of
TranslationMessageDataController (TMDC), moving TUPLE_SIZE=10 to TMDC,
adding segue to PrefsViewController (PVC) from main vc, adding done
button to PVC, with segue back to main, adding functions getNewLang and
getNewProj that can be overriden with new implementation of language
and project selection, removing mwapi (of commons-ios) from repository
as it is not related anymore. Preferred language and project are added
to TWUser, with default values of "en" and "!recent".