Adding CoreData support and functionality for ignoring rejected messages (CoreData integration doesn't yet work)
I tried to add support for CoreData, and I also added the required
functionality when tapping the X button in Proofread ViewController. I
also added in the addition of new messages to window (in
TranslationMessageDataController) a check, that checks if the received
messages weren't rejected by the current user, and loops until it gets
TUPLE_SIZE unrejected messages.

I changed the property userId of TWUser to NSInteger, because it
appears that the responses from TranslateWiki API give it as an integer.

I used Apple's tutorial for CoreData and I used this source to retrieve
code for CoreData support:
http://wiresareobsolete.com/wordpress/2009/12/adding-core-data-existing-
iphone-projects/