fixing both bugs: coredata and accept action
In the end, the bug was quite simple/annoying: the way I converted the
userid received from server - using (NSInteger)object - was wrong, and
probably saved the object id of the retreived NSNumber instead of its
value, or something similar. I switched userId back to NSString, and
now converted it properly to be NSString (before it was a strange
object named NSFCNumber or something).
So probably on each login request the userid changed, which explains
both the bug in accept action, and both the bug of the coredata (as I
search entries in the table according to both key and userid).
The coredata is now being saved properly between runs and after logout
and new login!