resign responder on done
+clearing "your translation…" on edit beginning

tried to do that also for background tap, but the background tap events
hid the didSelectRowAtIndexPath events. Maybe should be fixed
(currently activeTextView has no use - it was intended for background
tap). If commente code for background tap on MainViewController is
uncommented, then background tap will works but not selecting suggested
translations.